ipaccess: fix write path for OML/RSL message from BSC -> BTS
This patch fixes the write path for OML/RSL messages. I broke
this while trying to support the delivery of IPA CCM messages
using this path, which is not useful since they are directly
delivered by means of the file descriptor.

This patch fixes the corruption of messages leaving the BSC.