tests: use the virtual E1 line used by OML for RSL link
The ipaccess driver is tricky. Since it uses two ports for OML
and RSL links, the link for RSL needs to be done in the same
line that the OML link.

Now the BSC examples uses the virtual E1 line where the OML
link is for the RSL one.

This patch also verifies that the OML link is up before the
someone tries to get the RSL link up.