tests: Add a test stub for the subchan_demux tests
Tobias has worked on CSD support and Dieter has helped with a CSD
trau fix. This is a stub that should allow to easily write a test
case for the syncing issue observed by Tobias.

Run the tests:
$ make check

In case of failure:
$ vi tests/testsuite.dir/1/testsuite.log

Add the actual content:
$ edit tests/subchan_demux/subchan_demux_test.c

Update the test result:
$ ./tests/subchan_demux/subchan_demux_test > tests/subchan_demux/subchan_demux_test.ok