rtp: Randomize the initial sequence number for RFC compliance
The usage of random() is not that unpredictable but the closest
we can get right now.

RFC 3550:
The initial value of the sequence number SHOULD be random
(unpredictable) to make known-plaintext attacks on encryption
more difficult, even if the source itself does not encrypt
according to the method in Section 9.1, because the packets
may flow through a translator that does.  Techniques for
choosing unpredictable numbers are discussed in [17].

Fixes: SYS#403