unixsocket: Ensure superchannel is disabled for new connections
When the BSC is shut down or disconnected it may happen that the
SIU stays in superchannel mode, while the BTS is falling back
into its normal communication mode. This causes a permanent
link failure because the SIU is still trying to reconnect on
the superchannel.

This patch ensures that for every new connection the superchannel
is taken down by default. If the connection was still good, SIU
and BTS will switch back to normal mode together. If not. The
SIU will switch back alone while the BTS is already switched
back by itsself.