configure.ac: Fix RTP_SIGNAL_PTR_CAST config in FreeBSD build
Move function definition to the first argument of AC_LANG_PROGRAM
(prologue) instead of the second (body). Content of the body is writen
into the main() function of the file being compiled, while prologue is
writen before it.

Compiler from FreeBSD build host doesn't like definitions inside the
main function, it fails with the following error:
conftest.c:28:73: error: function definition is not allowed here

As the compilation test fails, unsigned int is used instead of void*, which
doesn't match the current headers in the current build, and then a
warning (or error with -Werror) is printed.

Change-Id: I83fc3997e5067b499bf43cea4ff953604d6cbeae