configure.ac: Fix condition check for RTP_SIGNAL_PTR_CAST define
While building libosmo-netif on a system with an old ortp (0.22), it
was seen that warning related to RTP_SIGNAL_PTR_CAST not being correctly
selected were being printed. The compilation didn't fail there as
expected because the issue was being handled as a warning instead of an
error, so the void* ptr case was always being selected.

Change-Id: Ib5e7569e68fdb921c63e0c12e20e73aafbebda46