ortp: enable scheduled mode on rtp socket creation
we use some functionalities of libortp that require to use rtp sessions
with the scheduled mode enabled. Since we do not enable the scheduled mode,
yet, we receive error messages in the log that complain about a non enabled
scheduled mode.

- Turn on scheduled mode using rtp_session_set_scheduling_mode() when
  osmo_rtp_socket_create() is called.

Note: this patch is related to chane:
I3a63c23f5ede47773b6a249a48ecebd5d3b45ace Unfortunately we observed a
segfault with ortp version 0.25.0 (ARM), which seems to be related to an
unintialized scheduling mode. We expect this patch to fix the problem.

Change-Id: I4b8f0349d5503bbe53745cc8903beb26f149f219
Closes: OS#3299