e1_input: refcount inc line during e1_sign_link_create, not during line update
Increase reference count when a new sign_link using the line is created.
Otherwise the line is freed too quickly during e1inp_sign_link_destroy()
with several TRX being used, since each RSL link is missing a reference.

That extra refcoutny in update() is not really needed given that we already
have a link assigned on the line when we call update.

Fixes: OS#3612
Fixes: OS#4094

Change-Id: I74405b0e87a89c17d58e87024f4aedbd30832013