trau_sync: prevent false positive synchronization on startup
The history buffer is filled with zeros on startup. Together with
incoming bits those may appear as a valid signal. (each trau frame
begins with 16 zeros). In order to prevent this lets set the history
buffer to all ones, so that it looks like the data in the history
buffer is from an unused timeslot.

Change-Id: Ie5cc7b9a1ab3d8940ef9d2147f5999e8d09ee40e
Related: OS#2547