Add missing build dependency to libosmocodec-dev
Since I0190872dd282bcfe0f97bb4f8ab8d09023f9f06b we are using a header
file provided by libosmocodec, which means we need to add it to our
dependencies (and include path).  Linking against the library is
not required, as we don't use any symbols at this point.

Change-Id: Ie4524165d8873f6c801e07d9f03c94f43c66110d