ipaccess: Fix wrong assertion in ipaccess_drop() when used by BTS code
The code wrongly assumed that ipaccess_drop was only called by BSC code,
which is wrong. ipaccess_drop is called by BTS run code path in __handle_ts1_write(),
if send() syscall fails (for instance because BSC becomes unreachable).
In that case, we need to account for the BTS role case which doesn't
store the line pointer into the ofd->data. In BTS case, it's a pointer
to the struct ipa_client_conn, which we leave up to sign_link_down() cb
to do whatever they please with.

Fixes: OS#4864
Change-Id: If763e5f7736921a4360ad9027ba075ef8e118934