ipaccess: do not block ipaccess_line_update on failure
The function ipaccess_line_update (called when the API user calls
e1inp_line_update and the line is of type ipaccess) internally sets
the line_already_initialized regardless whether the initalization is
successful or not.

This also means that if the initalization fails the API user is unable
to retry the initalition with corrected line options.

Change-Id: I90d330e2864f85b964628502afdb0c38eb54f15d
Related: SYS#4971