ipaccess: Allow reconfiguring the ipa line during line_update()
The e1inp_line_update() API is precisely aimed at that: tell that we
want to reconfigure some aspects of the e1 line. Hence, it makes no
sense to apply a no-op with the flag "line_already_initialized".
That flag was added a long time ago in order to avoid problems when
several BTS are attached to the same ipa server link. That case doesn't
apply to the BTS role, so we can simply recreate the ipa_cli_conn.

Fixes: d6216405b7363c94174a6d301b96f5ed4a8730ce
Change-Id: Iff8092f88726cf238dac2abb99e135ac5864272d