input/ipaccess: Don't start zero-ms timer on every write
Historically, before November 15, 2010 when commit
d49fc5ae24fc9d44d2b284392ab619cc7a69a876 was merged to [back then]
OpenBSC, before libosmo-abis became a separate library, we used to
have a 10us delay timer for subsequent writes to ip.access nanoBTS 900.

    ts: Reduce the delay to 0 for OML and RSL

    This is possible after not sending more than one OML command that
    requires an extra ACK. For the RSL line we do not need any speed
    limitation.

Ever since the above-mentioned commit, the BSC always sets that timeout
to zero, which makes libosmo-abis start a zero-microsecond libosmocore timer,
which in turn will make libosmocore call select/poll with zero timeout, which
makes the kernel return immediately.

Why not remove the timer completely? Because ipaccess-config.c still specifies
a non-zero signaling delay, and we cannot be sure that this is really not
needed.

So let's alter the code to only start the timer if it's non-zero

Change-Id: I9c379364e7e6afce35fc6316392b5b33748980f7