input/ipaccess: Remove unneeded osmo_fd_write_enable()
Recent commit optimize the same function by avoiding an extra poll loop
when e1i_ts->sign.delay was zero. Upon doing so, the
osmo_fd_write_disable() was moved to some conditional paths. Hence, the
WRITE flag is left set and we don't need to set it again in the code
path modified in this commit.

Fixes: 28fea7746bbc2fb8ca0f677a93559c0c9f4cff09
Change-Id: I84787b6de2a5ccc82bd8f19ce874e73708bc287f