trau_sync: add Ericsson RBS GPRS TRAU synchronization pattern (64kbps)
For GPRS TRAU frames (Ericsson calls them GSL frames), ericsson introduces
a propritary synchronization pattern. There are different patterns used
for synchronization, data frames with coding schemes CS1-4 and MCS1-8
share a slightly different pattern than the one that is used for
synchronization but the intersection between both patterns still
provides reliable synchronization. There is also a third pattern for
MCS9 that is not included in this patch.

Change-Id: I9d62089db557d37c924cc42fc8e3b608000ef4be
Related: OS#5198