Send raw data with e1d as it arrives from application
The application is responsible to send raw data in sync with the raw
data it receives. It also cares about sending data in advance, to
prevent underruns, if required. The length of data can vary.
(Now it is same procedure as with mISDN driver.)

OSMO_FD_WRITE flag must be disabled, if there is no message left
in the transmit queue.

Change-Id: I3e3ef623a02bffe75b168dc42514455b4df1007f