trau_pcu_ericsson: also support ER_UL_CHMOD_NB_UNKN in 16k mode.
With EGPRS Ericsson introduced an "UNKN" demodulation mode for normal
bursts. This would instruct the CCU to listen for GSMK and PSK
modulation simultaniously. For 16k there is only GSMK and there would be
no need for such a mode. Lets support it in the API anyway in order to
simplify API usage. When the caller selects it, we just use
ER_UL_CHMOD_NB_GMSK then.

Change-Id: I8492a6d042bd7354b5dc34069030305fe39a731f
Related: OS#5198