e1_input: add define constant for full subslot (whole E1 timeslot)
The struct gsm_e1_subslot has a memeber e1_ts_ss, which describes the
I.460 subslot to use inside an E1 timeslot. This is usually just an
integer number (e.g. 0,1,2,3 for a line with 4 16kbps subslots). However
it is also possible to use no subslots at all. This is encoded by
setting e1_ts_ss to 0xff. When this magic number is used somewhere in
the code it is not immediately obvious what it means, so lets add a
define constant.

Change-Id: I62725420d1cb2803b87fe5aa5e6ca1f3740be9eb
Related: OS#5198