trau_pcu_ericsson: change loglevel for CPS decoding
The CPS field is inside the MAC block. When we are unable to decode the
CPS field a message is printed on loglevel NOTICE. However, it is pretty
normal that the CCU often receives noise (always when no MS transmits)
so we see this error very often. Lets change the level to DEBUG.

Change-Id: I1f5adbf9da6be4fec2d8c93e781c5df62ef62238
Related: OS#5198