trau_pcu_ericsson_test: replace 64K CS1/CS2 testvectors
The CS1/CS2 testvectors for the 64K TRAU frame decoder are valid but the
contained MAC block consits of noise. At the time when the test data was
sampled it was not possible to get TRAU frames with valid content from
the CCU. This has changed, so lets update the testvectors with ones that
contain valid data.

Change-Id: I0b8aef1ce1eca5b10601a0a655f7e3b00915d3cc
Related: OS#5198