osmo_rtp2trau() for FR & EFR: correctly handle the no-data case
3GPP TS 48.060 sections 6.5.2 and 6.5.3 prescribe the behavior of
TRAU frame emitters under conditions of not having any speech or SID
frame to send.  In the case of osmo_rtp2trau() this no-data condition
is indicated with a zero rtp_len argument to the function, but this
case was not handled correctly: a garbage speech frame (previous
buffer content for the D bits in struct osmo_trau_frame) was being
generated.  The present change implements the rules of TS 48.060
sections 6.5.2 and 6.5.3: standard FR/EFR speech frame with BFI=1
for UL, special idle speech frame pattern for DL.

Change-Id: I7f21e2210bba9ef87f1af515a001a0cfc1c0a1d8