trau_pcu_ericsson: set uplink frame error bit correctly
At the moment we set the uplink frame error bit to 1 if ul_frame_err
is set to false. Unfortunately the er_gprs_trau_frame_encode pre set
all non T0 bits to 1 before calling the encoder function. This means
that with this logic the uplink frame error bit can never be set to 0.

Related: OS#6015
Change-Id: I74f56bc40ed074eb1588b202849870e2786ded24