e1d: get rid of strange file descriptor registered check
When the line update happens the list.next member of the file descriptor
is checked for NULL and != LLIST_POISON1. This directly tampers with the
llist API and might be problematic. Also we can tell by the file
descriptor number if the file descriptor is registered or not. An open
file descriptor is always registered.

Change-Id: I92c426271dc5455427471030fcf0749566e4c2ee
Related: OS#5983