build config: add --disable-ortp option
Having ortp as a mandatory dependency for Osmocom CNI suite is
very burdensome for users of OS distributions that don't provide
zero-effort readily installable packages for every piece of software
under the sun.  Furthermore, the actual usage of this dependency
in Osmocom is quite minimal: in a complete Osmocom 2G network,
only OsmoBTS uses the osmo_ortp layer provided by libosmo-abis,
and even that dependency may become optional in the future if we
implement an alternative jitter buffer mechanism natively in Osmocom.

Add --disable-ortp option to libosmo-abis configure that does what
it says, similar to the numerous --disable-* options in libosmocore.
When libosmo-abis is built with this option, osmo_ortp module is
excluded from libosmotrau, and <osmocom/trau/osmo_ortp.h> is omitted
from the installation.  Packages that require osmo_ortp will then
fail to build, but other Osmocom CNI components that require
libosmoabis and/or libosmotrau (but not osmo_ortp) will build
and work with ortp dependency eliminated.

Related: OS#6474
Change-Id: I554260483b94d812ac3a957c969a902870f53883