RTP->TRAU, FR/EFR: accept TW-TS-001 extended payloads
TW-TS-001 is an enhanced RTP transport format for FR and EFR
codecs that replicates the functionality and semantics of
GSM 08.60 TRAU-UL frames.  Given that libosmotrau functions
support converting from RTP to either TRAU-UL or TRAU-DL,
the option of TW-TS-001 fits in naturally.

It needs to be noted, however, that the fully correct path
in OsmoMGW from RTP input to E1 output will need to include
one more critical step _before_ the call to osmo_rtp2trau():
it will be a stateful preening engine that implements the rules
of TS 28.062 section C.3.2.1.1, a codec-dependent stateful
transform that can be fairly complex.  The necessary engines
will most likely be implemented in Themyscira codec libraries,
and OsmoMGW will need to be outfitted with an _optional_
dependency on those libs.  The "native" support for TW-TS-001
input in osmo_rtp2trau() for TRAU-DL (as opposed to TRAU-UL)
output is a fallback in the case of OsmoMGW built without any
Themyscira codec libraries.

Related: OS#6448
Change-Id: I3b2a33c1fcc5a7dafe06736dce5fa5b7b1138e54