trau_frame: fix 8k decoding
osmo_trau_frame_decode_8k() API was broken for most TRAU frame types:
the processing begins with identifying the sync pattern, but the check
for the leading 8 zeros in HRv1, AMR-low and AMR-6k7 formats was
broken in that it was checking for 16 zero bits instead of 8.

While at it, fix global namespace pollution: the bit8_0[] datum
used in this logic was defined as global when it should be static.

Change-Id: Idabc1283d477473b479f2d76d783ca9aeaf0af5d