trau2rtp unit test: add patterns that showcase BFI=1 SID=0
The previous patch added special handling for the corner case
in which the BTS indicates BFI=1 SID=0, but the bit pattern is
still SID by the bit counting rules of GSM 06.31 or 06.81
section 6.1.1.  The selection of test patterns in the unit test
already had some frames that exhibit this effect, taken from
actual Abis output of Nokia InSite BTS, but the exact effect
of our handling becomes more visible when the bit pattern forms
a perfect (not corrupted) SID frame, which I haven't been able
to produce with the InSite.  Therefore, add some hand-crafted
test frames that fully showcase the effect of our transform.

Change-Id: Ic6a0ad289864fe3e11192d158e3bbe221994eaaa