ipaccess: Fix wrong pointer passed in sign_link_up user callback
This bug was introduced in Jun 2011, 13 years ago, so probably this code
path is not (usually) used at all.

The commit introducing the regression changed sign_link_up() first
param from msgb to struct ipaccess_unit, but compiler didn't catch it
because it's being passed as a void pointer.

Fixes: c9c4fd39055ab02c0af59d5519e62edc456e3d0b
Related: OS#3622
Change-Id: I8e514360d5ed1c311af97f8fa728a2ff0a027522