e1_input: Store new msg in llist before calling driver->want_write()
This is needed for drivers handling the message synchronously during
want_write() op.
Until now this was not needed because probably all drivers are relying
on poll based write to tx the message, which meant using the msg
happened in a delayed form, asynchronously.

Once we move the ipaccess driver to use libosmo-netif's stream, we'll
need the msg to be available in the upper layer llist when want_write()
is called, because at that time it will pop it from that llist and
submit it to the lower layer stream tx_queue.

Change-Id: I2b4e7bed502964fe03e3dabaeed668713b4baa7a