input: Move ipaccess_bts_handle_ccm() ipaccess.c -> ipa.c
This function is no longer used in ipaccess.c, and actually relates to
IPA generic code which still exists in ipa.c, hence move it there.
Moreover, the function is declared in the "abis/ipa.h" public header,
not the "abis/ipaccess.h" header (which was already deprecated).

Once moved and it becomes clear this is not used in the e1_line ipaccess
driver and hence the "if (line) { ... ipa pong ... } code branch in the
function can be dropped, since it won't ever be executed.

Change-Id: I2378227515315e7fc0bd0844a1dc4f85e54b455a