input: Get rid of abis/ipa.h dependency in e1_input and ipaccess.c
abis/ipa.h will eventually be deprecated. The only current use of
abis/ipa.h in libosmo-abis itself is in helper code such as:
* tests/ipa_recv/ipa_recv_test.c
* src/ipa_proxy.c

ipaccess e1_input driver is also using struct ipa_keepalive_params,
which is only exported publicly in e1_input.h as a pointer and is only
really used internally, hence move a copy of the struct to internal.h
and use it internally (e1_input_vty.c) from there.

Leave a copy of the struct in its original abis/ipa.h (which as
mentioned contains code to be deprecated) for old users.

Change-Id: I4efb11ae44c28c4f56f71f5797966f6106493f2b